Update and rename watchdog_core.h to watchcat_core.h
